pyproject.toml
[project]
name = "scrabblebot-nosync"
version = "0.1.0"
description = "Mastodon bot that calculates Scrabble points for words in multiple languages"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "Mastodon.py>=1.8.0",
    "python-dotenv>=1.0.0",
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
